+++ bekawave-rust-backend/src/models/sales.rs
@@ -41,7 +41,7 @@ use sqlx::FromRow;
 /// - The `sales_id` field is initialized to `0` in the `build_sales` function and should be updated with a value assigned by the database after insertion.
 #[derive(Debug, Serialize, Deserialize, FromRow)]
 pub struct Sales {
-    pub sales_id: i32,
+    pub sales_id: i64,
     pub customer_id: i32,
     pub sales_pair_id: Option<i32>,
     pub sales_rep_id: Option<i32>,
